search

+++ webservices/normalTask_modifUser.php
@@ -23,14 +23,16 @@ function modifUser($tab){
 					$set .= "0,";
 				} else {
 					$set .= "1,";
-					$dest_fichier = $_SESSION['user'].'.'.get_extension($_FILES[$row[nom]][name]);
-					print $dest_fichier;
+					$ext =get_extension($_FILES[$row[nom]][name]);
+					if($ext==='jpeg') $ext='jpg';
+					$dest_fichier = $_SESSION['user'].'.'.$ext;
 					$photos=$_SERVER['DOCUMENT_ROOT']."/fliker/photos";
-					print $photos;
 					move_uploaded_file($_FILES['photo']['tmp_name'], $photos ."/". $dest_fichier);
-					$img = new ImageConverter($photos ."/". $dest_fichier,'jpg');
-					rename($_SESSION['user'].'.jpg',$photos ."/".$_SESSION['user'].'.jpg');
-					unlink($photos ."/". $dest_fichier);
+					if(!($ext==='jpg')){
+						$img = new ImageConverter($photos ."/". $dest_fichier,'jpg');
+						rename($_SESSION['user'].'.jpg',$photos ."/".$_SESSION['user'].'.jpg');
+						unlink($photos ."/". $dest_fichier);
+					}
 				}
